More conversions to BigInteger/Decimal

+++ src/main/java/com/enjin/enjincoin/sdk/model/service/identities/CreateIdentity.java
@@ -2,12 +2,13 @@ package com.enjin.enjincoin.sdk.model.service.identities;
 
 import com.enjin.enjincoin.sdk.graphql.GraphQLRequest;
 
+import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
 public class CreateIdentity extends GraphQLRequest.Builder {
 
-    public CreateIdentity withUserId(int userId) {
+    public CreateIdentity withUserId(BigInteger userId) {
         withParameter("user_id", userId);
         return this;
     }
