+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/CreateTradeData.java
@@ -3,6 +3,7 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Builder
@@ -18,6 +19,6 @@ public class CreateTradeData {
     private String secondPartyAddress;
 
     @SerializedName("second_party_identity_id")
-    private Integer secondPartyIdentityId;
+    private BigInteger secondPartyIdentityId;
 
 }
