+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/SetApprovalData.java
@@ -3,6 +3,7 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Builder
@@ -11,7 +12,7 @@ public class SetApprovalData {
     private String operator;
 
     @SerializedName("operator_id")
-    private Integer operatorId;
+    private BigInteger operatorId;
 
     private List<TokenIndexPairData> tokens;
