+++ src/main/java/com/enjin/enjincoin/sdk/model/service/requests/data/TransferData.java
@@ -3,18 +3,20 @@ package com.enjin.enjincoin.sdk.model.service.requests.data;
 import com.google.gson.annotations.SerializedName;
 import lombok.Builder;
 
+import java.math.BigInteger;
+
 @Builder
 public class TransferData {
 
     private String from;
 
     @SerializedName("from_id")
-    private Integer fromId;
+    private BigInteger fromId;
 
     private String to;
 
     @SerializedName("to_id")
-    private Integer toId;
+    private BigInteger toId;
 
     @SerializedName("token_id")
     private String tokenId;
